requirements.txt
omegaconf
rich
lz4
tqdm
hydra
tqdm
pyyaml
ipdb
submitit